Make store-worker reschedule itself
