+++ hazelcast/src/main/java/com/hazelcast/map/impl/mapstore/writebehind/StoreWorker.java
@@ -23,6 +23,8 @@ import com.hazelcast.map.impl.mapstore.MapDataStore;
 import com.hazelcast.map.impl.mapstore.MapStoreContext;
 import com.hazelcast.map.impl.mapstore.writebehind.entry.DelayedEntry;
 import com.hazelcast.map.impl.recordstore.RecordStore;
+import com.hazelcast.spi.ExecutionService;
+import com.hazelcast.spi.NodeEngine;
 import com.hazelcast.spi.partition.IPartition;
 import com.hazelcast.spi.partition.IPartitionService;
 import com.hazelcast.spi.properties.GroupProperty;
@@ -45,12 +47,14 @@ import static java.util.concurrent.TimeUnit.SECONDS;
  * <p/>
  * Only one {@link StoreWorker} task is created for a map on a member.
  */
-public class StoreWorker implements Runnable {
+public final class StoreWorker implements Runnable {
 
     private final String mapName;
     private final MapServiceContext mapServiceContext;
     private final IPartitionService partitionService;
+    private final ExecutionService executionService;
     private final WriteBehindProcessor writeBehindProcessor;
+    private final NodeEngine nodeEngine;
     private final long backupDelayMillis;
     private final long writeDelayMillis;
     private final int partitionCount;
@@ -62,11 +66,14 @@ public class StoreWorker implements Runnable {
      * @see #calculateHighestStoreTime
      */
     private long lastHighestStoreTime;
+    private boolean stopped;
 
     public StoreWorker(MapStoreContext mapStoreContext, WriteBehindProcessor writeBehindProcessor) {
         this.mapName = mapStoreContext.getMapName();
         this.mapServiceContext = mapStoreContext.getMapServiceContext();
-        this.partitionService = mapServiceContext.getNodeEngine().getPartitionService();
+        this.nodeEngine = mapServiceContext.getNodeEngine();
+        this.partitionService = nodeEngine.getPartitionService();
+        this.executionService = nodeEngine.getExecutionService();
         this.writeBehindProcessor = writeBehindProcessor;
         this.backupDelayMillis = getReplicaWaitTimeMillis();
         this.lastHighestStoreTime = Clock.currentTimeMillis();
@@ -74,9 +81,27 @@ public class StoreWorker implements Runnable {
         this.partitionCount = partitionService.getPartitionCount();
     }
 
+    public synchronized void start() {
+        stopped = false;
+        run();
+    }
+
+    public synchronized void stop() {
+        stopped = true;
+    }
 
     @Override
     public void run() {
+        try {
+            runInternal();
+        } finally {
+            if (!stopped) {
+                executionService.schedule(this, 1, SECONDS);
+            }
+        }
+    }
+
+    private void runInternal() {
         final long now = Clock.currentTimeMillis();
         // if this node is the owner of a partition, we use this criteria time.
         final long ownerHighestStoreTime = calculateHighestStoreTime(lastHighestStoreTime, now);
@@ -120,7 +145,6 @@ public class StoreWorker implements Runnable {
         }
 
         notifyFlush();
-
     }
 
     private static List<DelayedEntry> initListIfNull(List<DelayedEntry> list, int capacity) {
