--- hazelcast/src/main/java/com/hazelcast/map/impl/mapstore/writebehind/WriteBehindProcessors.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2008-2016, Hazelcast, Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hazelcast.map.impl.mapstore.writebehind;
-
-import com.hazelcast.map.impl.mapstore.MapStoreContext;
-
-/**
- * Static factory which creates a write behind processor.
- */
-public final class WriteBehindProcessors {
-
-    private WriteBehindProcessors() {
-    }
-
-    public static WriteBehindProcessor createWriteBehindProcessor(MapStoreContext mapStoreContext) {
-        return new DefaultWriteBehindProcessor(mapStoreContext);
-    }
-}
